refactor: move lógica de HomePageController para o estado de HomePage por não ser acessado em nenhum outro lugar
